requirements.txt
torch
transformers
accelerate
fastapi
uvicorn
flash-attn
pyannote.audio
torchaudio